chore(starknet_l1_provider_types): add crate
Only move, no logic changes, this is a setup for upcoming infra support.
