+++ Cargo.toml
@@ -40,6 +40,7 @@ members = [
   "crates/starknet_http_server",
   "crates/starknet_integration_tests",
   "crates/starknet_l1_provider",
+  "crates/starknet_l1_provider_types",
   "crates/starknet_mempool",
   "crates/starknet_mempool_p2p",
   "crates/starknet_mempool_p2p_types",
@@ -222,6 +223,7 @@ starknet_gateway = { path = "crates/starknet_gateway", version = "0.0.0" }
 starknet_gateway_types = { path = "crates/starknet_gateway_types", version = "0.0.0" }
 starknet_http_server = { path = "crates/starknet_http_server", version = "0.0.0" }
 starknet_l1_provider = { path = "crates/starknet_l1_provider", version = "0.0.0" }
+starknet_l1_provider_types = { path = "crates/starknet_l1_provider_types", version = "0.0.0" }
 starknet_mempool = { path = "crates/starknet_mempool", version = "0.0.0" }
 starknet_mempool_p2p = { path = "crates/starknet_mempool_p2p", version = "0.0.0" }
 starknet_mempool_p2p_types = { path = "crates/starknet_mempool_p2p_types", version = "0.0.0" }
