+++ crates/starknet_l1_provider/Cargo.toml
@@ -8,8 +8,8 @@ license.workspace = true
 [dependencies]
 indexmap.workspace = true
 papyrus_base_layer.workspace = true
-serde.workspace = true
 starknet_api.workspace = true
+starknet_l1_provider_types.workspace = true
 thiserror.workspace = true
 
 [dev-dependencies]
