+++ crates/starknet_l1_provider/src/l1_provider_tests.rs
@@ -3,11 +3,12 @@ use pretty_assertions::assert_eq;
 use starknet_api::test_utils::l1_handler::executable_l1_handler_tx;
 use starknet_api::transaction::TransactionHash;
 use starknet_api::{l1_handler_tx_args, tx_hash};
+use starknet_l1_provider_types::errors::L1ProviderError;
+use starknet_l1_provider_types::ValidationStatus;
 
-use crate::errors::L1ProviderError;
 use crate::test_utils::L1ProviderContentBuilder;
+use crate::L1Provider;
 use crate::ProviderState::{Pending, Propose, Uninitialized, Validate};
-use crate::{L1Provider, ValidationStatus};
 
 macro_rules! tx {
     (tx_hash: $tx_hash:expr) => {{
