+++ crates/starknet_l1_provider_types/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "starknet_l1_provider_types"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license.workspace = true
+
+[dependencies]
+serde.workspace = true
+thiserror.workspace = true
+
+[lints]
+workspace = true
